DDCYLS-169  - Added integration test package underneath test dir  - Added unit test package underneath test dir  - Moved Connector & Controller tests to the integration tests package  - Removed Connector & Controller tests from the unit tests package  - Removed a test which was testing if a localdate could be compared using cats syntax.
