+++ test/unit/uk/gov/hmrc/digitalservicestax/services/VolatilePersistenceSpec.scala
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package unit.uk.gov.hmrc.digitalservicestax.services
+
+import org.scalatest.flatspec.AnyFlatSpec
+import org.scalatest.matchers.should.Matchers
+
+class VolatilePersistenceSpec extends AnyFlatSpec with Matchers {
+
+  "randomDstNumber" should "not give duplicates" in {
+    val many = (1 to 1000).map{_ => VolatilePersistence.randomDstNumber}
+
+    many.distinct.size shouldBe (many.size)
+  }
+  
+}
